Add warning to custom-templates.yaml: Newton only
As per https://bugzilla.redhat.com/show_bug.cgi?id=1439223, this template for RH-OSP10j, 
was run in RH-OSP11 but the base definition of the compute role changed in 11. Added 
warning to top of custom-templates.yaml that this file is for 10 (Newton) only and suggested 
process to create a new version for those using 11 (Mitaka). 

FYI: for 11/Mitaka HCI will be productized and thus official documentation will be available, 
not an update RA.